move useSorted to file, fix subcategory type
