+++ src/redux/types.ts
@@ -1,3 +1,5 @@
+import { Subcategory } from "./subcategories/subcategoriesOperations";
+
 export interface BaseProduct {
   id: number;
   title: string;
@@ -7,17 +9,6 @@ export interface BaseProduct {
   image_url: string;
   amount: number;
   manufacturer_id: number;
-  subcategories: SubCategory[];
+  subcategories: Subcategory[];
   created_at: string;
 }
-
-export type SubCategory = {
-  id: number;
-  name: string;
-  value: string;
-  category_id: number | null;
-  group_id: number | null;
-  created_at: string;
-  updated_at: string;
-  deleted_at: string | null;
-};
